Add files via upload
What is a Recommender System:-
In simple words a recommender system is like having a really smart friend who knows exactly what we like and suggests things we might enjoy.

Imagine we are browsing Youtube, Netflix or Spotify. A recommender system looks at what we already watched or listened to before, as well as what other people with similar tastes have enjoyed. Then, it uses this information to recommend new movies, TV shows, songs, or products we might like.

The goal of a recommender system is to help us discover new things we'll love, making our experience more enjoyable and personalized. Here in this simple code we will see how to implement that. Lets see how to implement one.